pyproject.toml
[tool.poetry]
name = "tosti"
version = "4.0.0"
description = "Tartarus Order System for Take-away Items"
authors = ["Job Doesburg <job.doesburg@gmail.com>", "Lars van Rhijn <larsvanrhijn@gmail.com>"]
maintainers = ["Job Doesburg <job.doesburg@gmail.com>", "Lars van Rhijn <larsvanrhijn@gmail.com>"]
readme = "README.md"
repository = "https://github.com/KiOui/TOSTI"
license = "MIT"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1"
Pillow = "^10.4.0"
django-admin-autocomplete-filter = "^0.7.1"
django-import-export = "^3.3.7"
requests = "^2.32.0"
spotipy = "^2.25.1"
coverage = "^6.5.0"
django-guardian = "^2.4.0" # Does not support Django >= 4.0.
djangorestframework = "^3.15.2"
django-oauth-toolkit = "^2.4.0"
django-cors-headers = "^4.3.1"
uritemplate = "^4.1.1"
django-autocompletefilter = "^0.0.8" # No official support for Django >= 4.0.
django-bootstrap5 = "^23.4"
django-filter = "^24.1"
django-constance = {extras = ["database"], version = "^2.9.0"} # No official support for Django 5.0.
html2text = "^2020.1.16"
django-ical = "^1.9.2" # No official support for Django 5.0.
fontawesomefree = "^6.1.1"
django-tinymce = "^4.1.0"
bleach = { extras = ["css"], version = "^5.0.1" }
django-admin-rangefilter = "^0.12.1"
django-queryable-properties = "^1.9.2"
djangosaml2 = "^1.9.3" # No official support for Django 5.0.
django-model-utils = "^4.4.0"
inflection = "^0.5.1"

[tool.poetry.group.dev.dependencies]
black = "^24.3.0"
pydocstyle = "^6.1.1"
flake8 = "^6.0.0"
freezegun = "^1.2.1"
tomli = "^2.0.1"

[tool.poetry.group.prod.dependencies]
uwsgi = "^2.0.23"
psycopg2-binary = "^2.9.1"
sentry-sdk = "^2.8.0"

[tool.black]
exclude = '''
/(
  migrations
  | website/tosti/settings
)/
'''

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
